Modify bpf_redir to avoid redirecting FIN packets
When receiving SYN and ACK packets, the sk_skb eBPF program is not
executed, but when receiving FIN packets, it is executed.
The `skb->len` indicates the size of the packet payload, and
the `skb->len` of FIN packet is 0, so when skb->len is not 0,
the packet will be redirected to avoid redirecting FIN packets,
so that the server can receive FIN packet successfully.