Add a redirection method
The original method redirects the packets to the egress interface of
the server socket when the client and server are on the same computer
or on different computers. In fact, if the client and server are on the
same computer, the server socket can redirect the packet to the client
socket directly when it receives the packet.

Therefore, the new method will make different redirection operations
for the client and server are on the same computer and different
computers.